Refactor: 팀 드롭다운리스트 코드 분리 및 멤버 상태 스타일 변경
